Arrange packages to make all areas optional
- Split API into API and Providers
- Make OpenVPN and WireGuard optional in tests
