+++ Sources/API/ProviderMetadata+Codable.swift
