+++ Sources/Partout/API/APIV6Mapper.swift
@@ -23,6 +23,8 @@
 //  along with Partout.  If not, see <http://www.gnu.org/licenses/>.
 //
 
+#if canImport(PartoutAPI)
+
 import Foundation
 import PartoutAPI
 import PartoutCore
@@ -102,3 +104,5 @@ private extension API.V6.Mapper {
         }
     }
 }
+
+#endif
