+++ Sources/Providers/ProviderEntity.swift
