+++ Sources/Providers/ProviderHeuristic.swift
